requirements.txt
beautifulsoup4==4.11.1
django==3.2.9
django-bootstrap-pagination==1.7.0
django-cors-headers==3.10.0
django-crontab==0.7.1
django-extra-keywords==1.1.2
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn
feedparser==6.0.10
razdel==0.4.0
numpy==1.22
psycopg2-binary==2.8.6
pymorphy2==0.8
transliterate==1.10.2
PyJWT==2.4.0
pytz
openpyxl==3.1.2

